requirements.txt
# Core ML/DL libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Data handling
pandas>=1.3.0
openpyxl>=3.0.9

# Image processing
opencv-python>=4.5.0
Pillow>=8.0.0
albumentations>=1.3.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Model utilities
timm>=0.6.0

# Data download
gdown>=4.0.0

# Progress tracking (optional)
tqdm>=4.60.0

# Jupyter (if needed)
jupyter>=1.0.0
ipykernel>=6.0.0

# Development tools
black>=22.0.0
flake8>=4.0.0
pytest>=7.0.0
